refactor: Modularized Mission & Vision background
